reorganize the folder structure
- On Windows,
	app is not frameless,
	menu bar is auto hidden,
	close button is removed
