+++ src/components/sub-components/CDN/CDN.js
